controller was freezing
The controller was stopping when it lost wifi connection so an additional check added on the main loop to catch a dropped connection. Also noticed that the code for the mqtt connection was “blocking” which could also have been the cause of the freezing - ie going into an endless loop trying to reconnect when there was no wifi connection.